@@ -13,10 +13,15 @@ import { httpGetBinData } from "../../api/bin.api";
 
 const Header = (item) => {
   const redirectAPI = process.env.REACT_APP_BIN;
-
   const handleLogOut = async () => {
-    const isLogout = await httpGetBinData('/logout')
-    window.location.href = redirectAPI;
+  const cookies = document.cookie.split(';');
+  cookies.forEach(cookie => {
+    const cookieParts = cookie.split('=');
+    const cookieName = cookieParts[0].trim();
+    document.cookie = cookieName + '=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;';
+  });
+  localStorage.clear();
+     window.location.href = redirectAPI;
   };
   
   const items = [
@@ -40,7 +45,7 @@ const Header = (item) => {
     },
     {
       label: (
-        <a href="" style={{ display: "flex" }} onClick={handleLogOut}>
+        <a style={{ display: "flex" }} onClick={handleLogOut}>
           Sign out
         </a>
       ),
