@@ -1,3 +1,4 @@
+import { createDevice, deleteDevice, getDevice, updateDevice, resetLoadDataStatus, resetCreateDataStatus, resetUpdateDataStatus, resetDeleteDataStatus, } from "../../redux/reducer/deviceSlice";
 
 class DeviceData {
     getColumnShow() {
@@ -6,21 +7,29 @@ class DeviceData {
             title: 'Tên thiết bị',
             dataIndex: 'device_name',
             key: 'device_name',
+            type: 'text',
+            isEdit: true
         },
         {
             title: 'Loại thiết bị',
             dataIndex: 'device_type',
             key: 'device_type',
+            type: 'text',
+            isEdit: true
         },
         {
             title: 'Ngày bắt đầu',
             dataIndex: 'start_date',
             key: 'start_date',
+            type: 'date',
+            isEdit: true
         },
         {
             title: 'Ngày kết thúc',
             dataIndex: 'end_date',
             key: 'end_date',
+            type: 'text',
+            isEdit: true
           },
       ];
 
@@ -29,8 +38,21 @@ class DeviceData {
 
     getFunction() {
         return {
+            createData: createDevice,
+            getData: getDevice,
+            updateData: updateDevice,
+            deleteData: deleteDevice,
+            resetLoadDataStatus: resetLoadDataStatus,
+            resetCreateDataStatus: resetCreateDataStatus,
+            resetDeleteDataStatus: resetDeleteDataStatus,
+            resetUpdateDataStatus: resetUpdateDataStatus
         }
-      }
+    }
+
+    checkDateIndex(index) {
+        const dateIndex = ['start_date', 'end_date']
+        return dateIndex.includes(index);
+    }
 }
   
   
