@@ -1,4 +1,4 @@
-import { createData, deleteData, getData, updateData } from "../../redux/reducer/userSlice";
+import { createUser, deleteUser, getUser, resetLoadDataStatus, resetCreateDataStatus, resetUpdateDataStatus, resetDeleteDataStatus, updateUser } from "../../redux/reducer/userSlice";
 
 class UserData {
     getColumnShow() {
@@ -7,16 +7,25 @@ class UserData {
           title: 'Họ và tên',
           dataIndex: 'username',
           key: 'username',
+          isEdit: true
         },
         {
           title: 'Ngày sinh',
           dataIndex: 'birth',
           key: 'birth',
+          isEdit: true
         },
         {
           title: 'Số điện thoại',
           dataIndex: 'phone_number',
           key: 'phone_number',
+          isEdit: true
+        },
+        {
+          title: 'Thiết bị',
+          dataIndex: 'devices',
+          key: 'devices',
+          isEdit: false
         },
       ];
 
@@ -25,12 +34,21 @@ class UserData {
 
     getFunction() {
       return {
-        createData: createData,
-        getData: getData,
-        updateData: updateData,
-        deleteData: deleteData
+        createData: createUser,
+        getData: getUser,
+        updateData: updateUser,
+        deleteData: deleteUser,
+        resetLoadDataStatus: resetLoadDataStatus,
+        resetCreateDataStatus: resetCreateDataStatus,
+        resetDeleteDataStatus: resetDeleteDataStatus,
+        resetUpdateDataStatus: resetUpdateDataStatus
       }
     }
+
+    checkDateIndex(index) {
+      const dateIndex = ['birth']
+      return dateIndex.includes(index);
+  }
 }
   
   
