@@ -1,5 +1,10 @@
-import { httpDeleteData, httpGetData, httpPostData } from "../../api/common.api";
+import {
+  httpDeleteData,
+  httpGetData,
+  httpPostData,
+} from "../../api/common.api";
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
+import { getUserId } from "../../utils/storageUtils";
 
 export const loadStatus = {
   None: 0,
@@ -8,11 +13,11 @@ export const loadStatus = {
   Failed: 3,
 };
 
-export const getData = createAsyncThunk(
+export const getUser = createAsyncThunk(
   "/user",
   async (params, { rejectWithValue }) => {
     try {
-      const response = await httpGetData('/user');
+      const response = await httpGetData("/user");
       return response;
     } catch (error) {
       return rejectWithValue(
@@ -22,11 +27,11 @@ export const getData = createAsyncThunk(
   }
 );
 
-export const createData = createAsyncThunk(
+export const createUser = createAsyncThunk(
   "/create-user",
   async (params, { rejectWithValue }) => {
     try {
-      const response = await httpPostData('/user', params);
+      const response = await httpPostData("/user", params);
       return response;
     } catch (error) {
       return rejectWithValue(
@@ -36,7 +41,7 @@ export const createData = createAsyncThunk(
   }
 );
 
-export const updateData = createAsyncThunk(
+export const updateUser = createAsyncThunk(
   "/update-user",
   async (params, { rejectWithValue }) => {
     try {
@@ -50,7 +55,7 @@ export const updateData = createAsyncThunk(
   }
 );
 
-export const deleteData = createAsyncThunk(
+export const deleteUser = createAsyncThunk(
   "/delete",
   async (params, { rejectWithValue }) => {
     try {
@@ -64,72 +69,112 @@ export const deleteData = createAsyncThunk(
   }
 );
 
-const userSlice = createSlice({
-    name: 'user',
-    initialState: {
-      data: [],
-      loadDataStatus: loadStatus.None,
-      loadCreateDataStatus: loadStatus.None,
-      loadUpdateDataStatus: loadStatus.None,
-      loadDeleteDataStatus: loadStatus.None,
-    },
-    reducers: {
-      resetLoadDataStatus: (state, action) => {
-        state.data = [];
-        state.loadDataStatus = loadStatus.None;
-      },
-      resetCreateDataStatus: (state, action) => {
-        state.loadCreateDataStatus = loadStatus.None;
-      },
-      resetUpdateDataStatus: (state, action) => {
-        state.loadUpdateDataStatus = loadStatus.None;
-      },
-      resetDeleteDataStatus: (state, action) => {
-        state.loadDeleteDataStatus = loadStatus.None;
-      },
-    },
-    extraReducers: (builder) => {
-      builder
-        .addCase(getData.pending, (state, action) => {
-          state.loadDataStatus = loadStatus.Loading;
-        })
-        .addCase(getData.fulfilled, (state, action) => {
-          state.data = action.payload;
-          state.loadDataStatus = loadStatus.Success;
-        })
-        .addCase(getData.rejected, (state, action) => {
-          state.data = [];
-          state.loadDataStatus = loadStatus.Failed;
-        })
-        .addCase(createData.pending, (state, action) => {
-          state.loadCreateDataStatus = loadStatus.Loading;
-        })
-        .addCase(createData.fulfilled, (state, action) => {
-          state.loadCreateDataStatus = loadStatus.Success;
-        })
-        .addCase(createData.rejected, (state, action) => {
-          state.loadCreateDataStatus = loadStatus.Failed;
-        })
-        .addCase(updateData.pending, (state, action) => {
-          state.loadUpdateDataStatus = loadStatus.Loading;
-        })
-        .addCase(updateData.fulfilled, (state, action) => {
-          state.loadUpdateDataStatus = loadStatus.Success;
-        })
-        .addCase(updateData.rejected, (state, action) => {
-          state.loadUpdateDataStatus = loadStatus.Failed;
-        })
-        .addCase(deleteData.pending, (state, action) => {
-          state.loadDeleteDataStatus = loadStatus.Loading;
-        })
-        .addCase(deleteData.fulfilled, (state, action) => {
-          state.loadDeleteDataStatus = loadStatus.Success;
-        })
-        .addCase(deleteData.rejected, (state, action) => {
-          state.loadDeleteDataStatus = loadStatus.Failed;
-        })
+export const getUserById = createAsyncThunk(
+  "/user/id",
+  async (params, { rejectWithValue }) => {
+    try {
+      const user_id = getUserId();
+      console.log(user_id);
+      const response = await httpGetData(`/user/${user_id}`);
+      return response;
+    } catch (error) {
+      return rejectWithValue(
+        error?.response?.data?.message || error?.response || error
+      );
     }
+  }
+);
+
+const userSlice = createSlice({
+  name: "user",
+  initialState: {
+    data: [],
+    userData: {},
+    loadDataStatus: loadStatus.None,
+    loadCreateDataStatus: loadStatus.None,
+    loadUpdateDataStatus: loadStatus.None,
+    loadDeleteDataStatus: loadStatus.None,
+    loadUserDataStatus: loadStatus.None,
+  },
+  reducers: {
+    resetLoadDataStatus: (state, action) => {
+      state.data = [];
+      state.loadDataStatus = loadStatus.None;
+    },
+    resetCreateDataStatus: (state, action) => {
+      state.loadCreateDataStatus = loadStatus.None;
+    },
+    resetUpdateDataStatus: (state, action) => {
+      state.loadUpdateDataStatus = loadStatus.None;
+    },
+    resetDeleteDataStatus: (state, action) => {
+      state.loadDeleteDataStatus = loadStatus.None;
+    },
+    resetUserDataStatus: (state, action) => {
+      state.loadUserDataStatus = loadStatus.None;
+    },
+  },
+  extraReducers: (builder) => {
+    builder
+      .addCase(getUser.pending, (state, action) => {
+        state.loadDataStatus = loadStatus.Loading;
+      })
+      .addCase(getUser.fulfilled, (state, action) => {
+        state.data = action.payload;
+        state.loadDataStatus = loadStatus.Success;
+      })
+      .addCase(getUser.rejected, (state, action) => {
+        state.data = [];
+        state.loadDataStatus = loadStatus.Failed;
+      })
+      .addCase(createUser.pending, (state, action) => {
+        state.loadCreateDataStatus = loadStatus.Loading;
+      })
+      .addCase(createUser.fulfilled, (state, action) => {
+        state.loadCreateDataStatus = loadStatus.Success;
+      })
+      .addCase(createUser.rejected, (state, action) => {
+        state.loadCreateDataStatus = loadStatus.Failed;
+      })
+      .addCase(updateUser.pending, (state, action) => {
+        state.loadUpdateDataStatus = loadStatus.Loading;
+      })
+      .addCase(updateUser.fulfilled, (state, action) => {
+        state.loadUpdateDataStatus = loadStatus.Success;
+      })
+      .addCase(updateUser.rejected, (state, action) => {
+        state.loadUpdateDataStatus = loadStatus.Failed;
+      })
+      .addCase(deleteUser.pending, (state, action) => {
+        state.loadDeleteDataStatus = loadStatus.Loading;
+      })
+      .addCase(deleteUser.fulfilled, (state, action) => {
+        state.loadDeleteDataStatus = loadStatus.Success;
+      })
+      .addCase(deleteUser.rejected, (state, action) => {
+        state.loadDeleteDataStatus = loadStatus.Failed;
+      })
+      .addCase(getUserById.pending, (state, action) => {
+        state.loadUserDataStatus = loadStatus.Loading;
+      })
+      .addCase(getUserById.fulfilled, (state, action) => {
+        state.loadUserDataStatus = loadStatus.Success;
+        state.userData = action.payload;
+        console.log(action.payload);
+      })
+      .addCase(getUserById.rejected, (state, action) => {
+        state.loadUserDataStatus = loadStatus.Failed;
+        state.userData = {};
+      });
+  },
 });
 
 const { reducer: userReducer } = userSlice;
+export const {
+  resetLoadDataStatus,
+  resetCreateDataStatus,
+  resetUpdateDataStatus,
+  resetDeleteDataStatus,
+  resetUserDataStatus,
+} = userSlice.actions;
 export default userReducer;
