@@ -1,18 +1,18 @@
 import axios from 'axios';
-
+import { getCookie, getLocalStorage } from './storageUtils';
 export const axiosMethod = {
   GET: 'GET',
   POST: 'POST',
   PUT: 'PUT',
   DELETE: 'DELETE'
 };
-
-export const axiosRequest = (url, method, token, data) => {
+export const axiosRequest = (url, method, data) => {
   const axiosConfig = {
     url,
     method,
     headers: {}
   };
+  const token = window.hasCookie.access_token;
   if (method !== axiosMethod.GET) {
     axiosConfig.headers['Content-Type'] = `application/json`;
   }
@@ -20,6 +20,6 @@ export const axiosRequest = (url, method, token, data) => {
     axiosConfig.headers['authorization'] = `Bearer ${token}`;
     axiosConfig.headers['X-XSRF-TOKEN'] = token;
   }
-  axiosConfig.data = data;
+  if(data) axiosConfig.data = data;
   return axios(axiosConfig);
 };
