@@ -1,4 +1,7 @@
+const { response } = require("express");
 const RecordService = require("../services/RecordService");
+const fileUploader = require("../services/FileService");
+
 class RecordController {
   async getAll(req, res, next) {
     console.log(`[P]:::Get all records: `);
@@ -14,7 +17,7 @@ class RecordController {
     try {
       await RecordService.add(req.body);
       return res.status(200).json({
-        message: "Create records successful!",
+        message: "Create record successful!",
       });
     } catch (err) {
       return res.status(400).json({
@@ -26,14 +29,18 @@ class RecordController {
   async getRecordById(req, res, next) {
     console.log(`[P]:::Get record by id: `, req.params.recordId);
     const record = await RecordService.getRecordById(req.params.recordId);
-    if (record) {
-      return res.status(200).json({
-        message: "Get records by id successful!",
-        metadata: record,
-      });
-    }
-    return res.status(404).json({
-      message: "Record not existed!",
+    return res.status(200).json({
+      message: "Get records by id successful!",
+      metadata: record,
+    });
+  }
+
+  async getDataRecord(req, res, next) {
+    console.log(`[P]:::Get data record: `, req.params.length);
+    const data = await RecordService.getDataRecord(req.params.length);
+    return res.status(200).json({
+      message: "Get data record successful!",
+      metadata: data,
     });
   }
 
@@ -42,27 +49,49 @@ class RecordController {
     const recordByDevice = await RecordService.getRecordByDeviceId(
       req.params.deviceId
     );
-    if (recordByDevice) {
-      return res.status(200).json({
-        message: "Get records by device id successful!",
-        metadata: recordByDevice,
-      });
-    }
-    return res.status(404).json({
-      message: "Record not existed!",
+    return res.status(200).json({
+      message: "Get records by device id successful!",
+      metadata: recordByDevice,
+    });
+  }
+
+  async getRecordByUserId(req, res, next) {
+    console.log(`[P]:::Get record by user id: `, req.params.userId);
+    const recordByUser = await RecordService.getRecordByUserId(
+      req.params.userId
+    );
+    return res.status(200).json({
+      message: "Get records by user id successful!",
+      metadata: recordByUser,
+    });
+  }
+
+  async getRecordByStartTime(req, res, next) {
+    console.log(`[P]:::Get record by start time: `, req.params.time);
+    const recordByStartTime = await RecordService.getRecordByStartTime(
+      req.params.time
+    );
+    return res.status(200).json({
+      message: "Get records by start time successful!",
+      metadata: recordByStartTime,
+    });
+  }
+
+  async getRecordByEndTime(req, res, next) {
+    console.log(`[P]:::Get record by end time: `, req.params.time);
+    const recordByEndTime = await RecordService.getRecordByEndTime(
+      req.params.time
+    );
+    return res.status(200).json({
+      message: "Get records by end time successful!",
+      metadata: recordByEndTime,
     });
   }
 
   async updateRecordById(req, res, next) {
-    console.log(`[P]:::Update record by id: `, req.params.recordId, req.body);
-    const id = req.params.recordId;
-    const checkRecord = await RecordService.getRecordById(id);
-    if (!checkRecord) {
-      return res.status(404).json({
-        message: "Id not found!",
-      });
-    }
-    await RecordService.updateRecordById(id)
+    console.log(`[P]:::Update record by id: `, req.body.id);
+    const id = req.body.id;
+    await RecordService.updateRecordById(req.body, id)
       .then(() => {
         return res.status(200).json({
           message: "Update record by id successful!",
@@ -77,14 +106,8 @@ class RecordController {
   }
 
   async deleteRecordById(req, res, next) {
-    console.log(`[P]:::Delete record by id: `, req.body.id);
-    const id = req.body.id;
-    const checkRecord = await RecordService.getRecordById(id);
-    if (!checkRecord) {
-      return res.status(400).json({
-        message: "Id not found!",
-      });
-    }
+    console.log(`[P]:::Delete record by id: `, req.params.id);
+    const id = req.params.id;
     await RecordService.deleteRecordById(id)
       .then(() => {
         return res.status(200).json({
@@ -96,6 +119,14 @@ class RecordController {
           message: "Error when delete record!",
         });
       });
+  }
+  
+  async UploadFileRecord(req, res, next) {
+    try {
+        await RecordService.uploadFileRecord(req, res, next);
+      } catch (err) {
+        console.log(err);
+      }
   }
 }
 
