@@ -12,11 +12,35 @@ class RecordRepository {
     });
   }
 
-  async getRecordsByDeviceId(id) {
+  async getRecordByDeviceId(id) {
     return await RecordDTO.findAll({
       where: {
         device_id: id,
       },
+    });
+  }
+
+  async getRecordByUserId(id) {
+    return await RecordDTO.findAll({
+      where: {
+        user_id: id,
+      },
+    });
+  }
+
+  async getRecordByStartTime(time) {
+    return await RecordDTO.findAll({
+      where: {
+        start_time: time,
+      }
+    });
+  }
+
+  async getRecordByEndTime(time) {
+    return await RecordDTO.findAll({
+      where: {
+        end_time: time,
+      }
     });
   }
 
